mod install.sh to recognise WSL

+++ install.sh
@@ -3,6 +3,8 @@
 # ターミナル(OS)の種類を判断
 if [ "$(uname)" == 'Darwin' ]; then
   OS='Mac'
+elif [ $(echo $(uname -r) | grep -e 'Microsoft') ]; then
+  OS='WSL'
 elif [ "$(expr substr $(uname -s) 1 5)" == 'Linux' ]; then
   OS='Linux'
 elif [ "$(expr substr $(uname -s) 1 10)" == 'MINGW32_NT' ]; then
@@ -11,19 +13,22 @@ else
   echo "Your platform ($(uname -a)) is not supported."
   OS='undefined' 
 fi
+echo $OS' terminal'
 
 # Mac の場合，Mac ディレクトリを参照
-for f in Mac/.??*
-do
+if [ $OS -e "Mac" ]; then
+  for f in Mac/.??*
+  do
     [[ "$f" == ".DS_Store" ]] && continue
     ln -s $HOME/dotfiles/$f $HOME/${f#*/}
-done
+  done
+fi
 
 # 各dotfileにリンクを貼る
 for f in .??*
 do
-    [[ "$f" == ".git" ]] && continue
-    [[ "$f" == ".gitignore" ]] && continue
-    [[ "$f" == ".DS_Store" ]] && continue
-    ln -s $HOME/dotfiles/$f $HOME/$f
+  [[ "$f" == ".git" ]] && continue
+  [[ "$f" == ".gitignore" ]] && continue
+  [[ "$f" == ".DS_Store" ]] && continue
+  ln -s $HOME/dotfiles/$f $HOME/$f
 done
